Changed line 106 of 'analysis' module.
When passing a risk-free rate DataFrame to modify returns to be excess returns, a `ValueError` was being thrown because the column names of the risk-free rate DataFrame and the returns columns did not align. This was corrected by passing `rf.values` to the `returns.sub()` function on line 106 of the `analysis` module.